requirements.txt
appdirs==1.4.0
boto==2.46.1
dj-database-url==0.4.1
Django==1.10.5
django-angular==0.8.4
django-cors-headers==2.0.2
django-filter==1.0.1
django-smtp-ssl==1.0
django-storages==1.5.2
djangorestframework==3.5.3
gunicorn==19.6.0
httplib2==0.10.3
Jinja2==2.9.5
Markdown==2.6.8
MarkupSafe==0.23
npm==0.1.1
oauth2client==4.0.0
optional-django==0.1.0
packaging==16.8
psycopg2==2.6.2
pyasn1==0.2.2
pyasn1-modules==0.0.8
pyparsing==2.1.10
rsa==3.4.2
six==1.10.0
uritemplate==3.0.0
whitenoise==3.3.0
